requirements.txt
azure-iot-device>=2.13.0
azure-identity>=1.15.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
